+++ config/serviceproviders.php
@@ -10,7 +10,6 @@
 
 return [
     'enabled' => [
-        'App\\Providers\\AppServiceProvider',
         'App\\Providers\\LemmaServiceProvider',
         // 'App\\Providers\\EventServiceProvider',
     ],